requirements.txt
Flask==2.3.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
praw==7.7.1
textblob==0.17.1
nltk==3.8.1
Jinja2==3.1.3
python-dotenv==1.0.1
